streams: fix readable high water mark measured in characters
The current implementation of the readable stream converts
the chunk to a string, before the check is made to compare
the chunk length to the highWaterMark. This is incongruent
with the stream documentation.

Additionally, tests are added to check this at a lower
level in utf8 using characters that do not have the same
character length as its byte length.